Former-commit-id: 32e6caba69cdf93f2563ec981b194fb60408e6a5

+++ src/mongo/db/curop.cpp
@@ -283,7 +283,7 @@ void CurOp::reportCurrentOpForClient(OperationContext* opCtx,
     }
 
     if (clientOpCtx) {
-        infoBuilder->append("opid", clientOpCtx->getOpID());
+        infoBuilder->append("opid", static_cast<int>(clientOpCtx->getOpID()));
         if (clientOpCtx->isKillPending()) {
             infoBuilder->append("killPending", true);
         }
